'Found locally' implementation
This change improves how availability from IncDocs is handled and
considers the possibility of an article being availabily locally i.e. at
the requesting library.
For this use case, a request to IncDocs is never placed. Instead, the
contentLocation and fullTextFile are added to the request and an email
is queued to the requesting patron.
The template notice used for this email must be configured in the plugin
configuration. See README.md for a template example.